requirements.txt
django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
google-generativeai==0.4.0
Pillow>=10.2.0
python-magic==0.4.27
gunicorn==21.2.0